requirements.txt
langchain
langchain_community
sentence-transformers
faiss-cpu
transformers
fastapi
uvicorn
scikit-learn
streamlit
